+++ src/pages/corp/api/actions.ts
@@ -1,7 +1,7 @@
 import type { APIRoute } from "astro";
 import type { InferSelectModel } from "drizzle-orm";
-import { jsonError, jsonResponse, JsonResponseError } from "../../../server/responses";
 import { Action } from "../../../database/schema";
+import { jsonError, jsonResponse, JsonResponseError } from "../../../server/responses";
 import { postJsonToBot, requireCorpMember } from "../../../server/submissions";
 
 export const prerender = false;
