+++ src/pages/corp/api/motions.ts
@@ -1,7 +1,7 @@
 import type { APIRoute } from "astro";
 import type { InferSelectModel } from "drizzle-orm";
-import { jsonError, jsonResponse, JsonResponseError } from "../../../server/responses";
 import { Motion } from "../../../database/schema";
+import { jsonError, jsonResponse, JsonResponseError } from "../../../server/responses";
 import { postFormDataToBot, requireCorpMember } from "../../../server/submissions";
 
 export const prerender = false;
