+++ src/pages/corp/api/sights.ts
@@ -1,7 +1,7 @@
 import type { APIRoute } from "astro";
 import type { InferSelectModel } from "drizzle-orm";
-import { jsonError, jsonResponse, JsonResponseError } from "../../../server/responses";
 import { Sight } from "../../../database/schema";
+import { jsonError, jsonResponse, JsonResponseError } from "../../../server/responses";
 import { postFormDataToBot, requireCorpMember } from "../../../server/submissions";
 
 export const prerender = false;
