+++ src/server/submissions.ts
@@ -1,16 +1,22 @@
 import { fetchWithHmac } from "@latex4000/fetch-hmac";
-import { submissionConfig } from "./config";
-import { JsonResponseError } from "../responses";
+import type { APIContext } from "astro";
+import { eq } from "drizzle-orm";
+import db from "../database/db";
+import { Member } from "../database/schema";
+import { JsonResponseError } from "./responses";
+
+const botBaseUrl = process.env.SUBMISSION_BOT_BASE_URL ?? "http://localhost:5556";
+const botHmacKey = process.env.SECRET_HMAC_KEY ?? null;
 
 function getBotBaseUrl(): string {
-    const baseUrl = submissionConfig.botBaseUrl;
+    const baseUrl = botBaseUrl;
     if (!baseUrl)
         throw new JsonResponseError("Submission bot endpoint is not configured", 500);
     return baseUrl;
 }
 
 function getBotHmacKey(): string {
-    const secret = submissionConfig.botHmacKey;
+    const secret = botHmacKey;
     if (!secret)
         throw new JsonResponseError("SECRET_HMAC_KEY is not configured", 500);
     return secret;
@@ -60,4 +66,24 @@ export function postJsonToBot<T>(path: string, payload: unknown): Promise<T> {
 
 export function postFormDataToBot<T>(path: string, formData: FormData): Promise<T> {
     return requestBot(path, formData);
+}
+
+export async function requireCorpMember(context: APIContext) {
+    const discord = context.locals.session.data.memberDiscord;
+    if (!discord)
+        throw new JsonResponseError("Not authenticated", 401);
+
+    const member = await db
+        .select()
+        .from(Member)
+        .where(eq(Member.discord, discord))
+        .get();
+
+    if (!member)
+        throw new JsonResponseError("Member does not exist", 403);
+
+    if (member.deleted)
+        throw new JsonResponseError("Member has been deleted", 403);
+
+    return member;
 }
