--- src/server/submissions/index.ts
@@ -1,3 +1,0 @@
-export * from "./config";
-export * from "./member";
-export * from "./botApi";
